refactor(ast): organize visitor
