+++ crates/ast/src/lib.rs
@@ -2,6 +2,7 @@ pub mod ast;
 pub mod evaluator;
 pub mod printer;
 pub mod validator;
+pub mod visitor;
 
 pub use ast::{
     ASTBinaryExpression, ASTBinaryOperator, ASTBinaryOperatorKind, ASTExpression, ASTLetStatement,
@@ -11,3 +12,4 @@ pub use ast::{
 pub use evaluator::ASTEvaluator;
 pub use printer::ASTPrinter;
 pub use validator::SymbolChecker;
+pub use visitor::ASTVisitor;
