+++ crates/ast/src/visitor.rs
@@ -1,0 +1,70 @@
+use crate::ast::{
+    ASTBinaryExpression, ASTExpression, ASTExpressionKind, ASTLetStatement, ASTNumberExpression,
+    ASTParenthesizedExpression, ASTStatement, ASTStatementKind, ASTUnaryExpression,
+    ASTVariableExpression,
+};
+use text::span::TextSpan;
+
+pub trait ASTVisitor<'de> {
+    fn do_visit_statement(&mut self, statement: &ASTStatement<'de>) {
+        match &statement.kind {
+            ASTStatementKind::Expression(expr) => {
+                self.visit_expression(expr);
+            }
+            ASTStatementKind::LetStatement(expr) => {
+                self.visit_let_statement(expr);
+            }
+        }
+    }
+
+    fn visit_let_statement(&mut self, let_statement: &ASTLetStatement<'de>);
+
+    fn visit_statement(&mut self, statement: &ASTStatement<'de>) {
+        self.do_visit_statement(statement);
+    }
+
+    fn do_visit_expression(&mut self, expression: &ASTExpression<'de>) {
+        match &expression.kind {
+            ASTExpressionKind::NumberLiteral(number) => {
+                self.visit_number_expression(number);
+            }
+            ASTExpressionKind::BinaryExpression(expr) => {
+                self.visit_binary_expression(expr);
+            }
+            ASTExpressionKind::ParenthesizedExpression(expr) => {
+                self.visit_parenthesized_expression(expr);
+            }
+            ASTExpressionKind::Error(span) => {
+                self.visit_error(span);
+            }
+            ASTExpressionKind::Variable(expr) => {
+                self.visit_variable_expression(expr);
+            }
+            ASTExpressionKind::UnaryExpression(expr) => self.visit_unary_expression(expr),
+        }
+    }
+
+    fn visit_variable_expression(&mut self, variable_expression: &ASTVariableExpression<'de>);
+
+    fn visit_expression(&mut self, expression: &ASTExpression<'de>) {
+        self.do_visit_expression(expression);
+    }
+
+    fn visit_number_expression(&mut self, number: &ASTNumberExpression);
+
+    fn visit_error(&mut self, span: &TextSpan);
+
+    fn visit_unary_expression(&mut self, unary_expression: &ASTUnaryExpression<'de>);
+
+    fn visit_binary_expression(&mut self, binary_expression: &ASTBinaryExpression<'de>) {
+        self.visit_expression(&binary_expression.left);
+        self.visit_expression(&binary_expression.right);
+    }
+
+    fn visit_parenthesized_expression(
+        &mut self,
+        parenthesized_expression: &ASTParenthesizedExpression<'de>,
+    ) {
+        self.visit_expression(&parenthesized_expression.expression);
+    }
+}
